fix(precrawl): run on api pod instead of over bullmq
