+++ apps/api/src/index.ts
@@ -10,7 +10,6 @@ import {
   getGenerateLlmsTxtQueue,
   getDeepResearchQueue,
   getBillingQueue,
-  getPrecrawlQueue,
 } from "./services/queue-service";
 import { v0Router } from "./routes/v0";
 import os from "os";
@@ -83,7 +82,6 @@ const { addQueue, removeQueue, setQueues, replaceQueues } = createBullBoard({
     new BullMQAdapter(getGenerateLlmsTxtQueue()),
     new BullMQAdapter(getDeepResearchQueue()),
     new BullMQAdapter(getBillingQueue()),
-    new BullMQAdapter(getPrecrawlQueue()),
   ],
   serverAdapter: serverAdapter,
 });
